refactor: 폴더 이동 및 파일 이름 변경
[#823]
