+++ frontend/src/components/common/Text/Text.tsx
@@ -2,7 +2,7 @@ import type { CSSProp } from 'styled-components';
 
 import type { HTMLAttributes } from 'react';
 
-import type { SpacingProps } from '@common/systems';
+import type { SpacingProps } from '@common/styles/spacing';
 
 import { StyledText } from './Text.style';
